split some actions
